Scales are now defined at first frame
