+++ src/h5web/visualizations/shared/models.ts
@@ -1,3 +1,5 @@
+import { ScaleLinear } from 'd3-scale';
+
 export type Domain = [number, number];
 
 export type Size = { width: number; height: number };
@@ -7,4 +9,9 @@ export type AxisOffsets = { left: number; bottom: number };
 export interface AxisDomains {
   left: Domain;
   bottom: Domain;
+}
+
+export interface TwoDimScale {
+  x: ScaleLinear<number, number>;
+  y: ScaleLinear<number, number>;
 }
